pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "learnagent"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "openai",
  "python-dotenv",
  "pydantic",
  "tenacity",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
